gtkdnd: use css class for drag highlight
Instead of a temporary style class and a caire rectangle around
the widget.

The only drawback about this is that GtkTreeView applies all the
style to the widget itself to its cells, so when gtk_drag_highlight
is applied, the cells show borders.
To prevent that, remove the dnd style class from the draw function
on GtkTreeView. The drawback about this solution is that GtkTreeView
won't draw a border to indicate "general" drop target availability, but
individual cells will still work as expect for highlight since they are
drawn in an internal way in GtkTreeView.

https://bugzilla.gnome.org/show_bug.cgi?id=749910